qoqo_for_braket_devices/pyproject.toml
[project]
name = "qoqo_for_braket_devices"
version = "0.1.0"
dependencies = [
  'qoqo_calculator_pyo3>=1.1',
  'qoqo>=1.5',
  'aws-braket-sdk',
]
license = {text="Apache-2.0 AND Apache-2.0 with LLVM-exception AND MIT AND Unicode-DFS-2016 AND BSD-2-Clause AND BSD-3-CLause"}
maintainers = [{name = "HQS Quantum Simulations GmbH", email = "info@quantumsimulations.de"}]
requires-python = ">=3.8"

[project.optional-dependencies]
docs = [
    'numpy',
    'sphinx>=2.1',
    'nbsphinx',
    'pygments',
    'recommonmark',
    'myst_parser',
    'sphinx_rtd_theme',
    'tomli'
]

[build-system]
requires = ["maturin>=0.14,<0.15"]
build-backend = "maturin"

[tool.maturin]
bindings = "pyo3"
compatibility = "manylinux2014"
skip-auditwheel = false
strip = true
profile = "release"
